Change structure

+++ src/main/java/imss/gob/mx/cohorte/services/almacenamiento/refrigerador/PisoRefrigeradorService.java
@@ -2,7 +2,6 @@ package imss.gob.mx.cohorte.services.almacenamiento.refrigerador;
 
 import imss.gob.mx.cohorte.modules.almacenamiento.refrigerador.PisoRefrigerador;
 import imss.gob.mx.cohorte.modules.almacenamiento.refrigerador.PisoRefrigeradorRepository;
-import imss.gob.mx.cohorte.modules.almacenamiento.refrigerador.RefrigeradorRepository;
 import imss.gob.mx.cohorte.utils.Exceptions.exceptions.ObjNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -70,8 +69,11 @@ public class PisoRefrigeradorService {
         return pisoRefrigeradorRepository.save(pisoBD);
     }
 
+    @Transactional
     public void deletePiso(Long id) {
-        // Implementación básica para eliminar un piso
+        if (!pisoRefrigeradorRepository.existsById(id)) {
+            throw new ObjNotFoundException("No se encontró el piso con id: " + id);
+        }
         pisoRefrigeradorRepository.deleteById(id);
     }
 }
